% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_functions.R
\name{add_to_g}
\alias{add_to_g}
\title{Initialize a mixture of multivariate normals}
\usage{
add_to_g(data, cov_methods, grid, g = NULL)
}
\arguments{
\item{data}{a mash data object, e.g. as created by \code{set_mash_data}}

\item{cov_methods}{methods to use to create covariance matrices; passed to \code{cov_methods}}

\item{grid}{a vector of scalars that are to be used to scale the covariance matrices}

\item{g}{optionally a previously-inititalized g; in this case the new g adds new covariances to this g}
}
\value{
a list with two elements, the covariance matrices and their mixture proportions
As well as adding new covariance matrices to g, the mixture proportions are re-initialized with \code{initialize_pi}.
}
\description{
Initialize a mixture of multivariate normals
}

