% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_functions.R
\name{autoselect_grid}
\alias{autoselect_grid}
\title{Automatically select
TODO: this is just a place-holder}
\usage{
autoselect_grid(data)
}
\description{
Automatically select
TODO: this is just a place-holder
}

