% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed.R
\name{bovy_wrapper}
\alias{bovy_wrapper}
\title{Fit extreme deconvolution to mash data using Bovy et al 2011}
\usage{
bovy_wrapper(data, Ulist_init, subset = NULL, ...)
}
\arguments{
\item{data}{mash data object}

\item{Ulist_init}{a list of covariance matrices to initialize to}

\item{subset}{the indices of the observations to be used (defaults
to all of them)}

\item{...}{arguments to be passed to \code{extreme_deconvolution}
function, such as \code{tol}, \code{maxiter}.}
}
\value{
the fitted mixture: a list of mixture proportions and
covariance matrices
}
\description{
This is an internal (non-exported) function. This help
  page provides additional documentation mainly intended for
  developers and expert users.
}
\details{
This is a wrapper to ExtremeDeconvolution::extreme_deconvolution
It fixes the projection to be the identity, and the means to be 0
}
\keyword{internal}
