% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{calc_lik_matrix}
\alias{calc_lik_matrix}
\title{Compute matrix of conditional likelihoods.}
\usage{
calc_lik_matrix(data, Ulist, log = FALSE, algorithm.version = c("R",
  "Rcpp"))
}
\arguments{
\item{data}{A \code{mash} data object; e.g., created by
\code{\link{set_mash_data}}.}

\item{Ulist}{List containing the prior covariance matrices.}

\item{log}{If \code{TRUE}, the return value is a matrix of log-
likelihoods.}

\item{algorithm.version}{Explain what this argument is for.}
}
\value{
J x P matrix of multivariate normal likelihoods, p(bhat |
    Ulist[p], V).
}
\description{
computes matrix of condition likelihoods for each of J
    rows of Bhat for each of P prior covariances.
}
