% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{calc_lik_vector}
\alias{calc_lik_vector}
\title{calc_lik_vector}
\usage{
calc_lik_vector(bhat, V, Ulist, log = FALSE)
}
\arguments{
\item{bhat}{Rx1 vector of bhat}

\item{V}{RxR covariance matrix for likelihood}

\item{Ulist}{list of prior covariance matrices}

\item{log}{if true computes log-likelihood}
}
\value{
P vector of multivariate normal likelihoods, with pth element p(bhat | Ulist[p], V)
}
\details{
computes vector of likelihoods for bhat for each of P prior covariances
}
