% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{calc_relative_lik_matrix}
\alias{calc_relative_lik_matrix}
\title{Calculate matrix of relative likelihoods (likelihoods, normalized to avoid numeric issues)}
\usage{
calc_relative_lik_matrix(Bhat, Shat, Ulist)
}
\arguments{
\item{Bhat}{J x R matrix of betahat values}

\item{Shat}{J x R matrix of standard errors}

\item{Ulist}{list of prior covariance matrices}
}
\value{
J x P matrix of likelihoods, p(bhat[j] | Ulist[p], V), but normalized so that the max in each row is 1
}
\details{
computes matrix of relative likelihoods for each of J rows of Bhat for each of P prior covariances
}

