% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mash.R
\name{compute_alt_loglik_from_matrix_and_pi}
\alias{compute_alt_loglik_from_matrix_and_pi}
\title{Compute vector of alternative loglikelihoods from a matrix of log-likelihoods and fitted pi}
\usage{
compute_alt_loglik_from_matrix_and_pi(pi_s, lm)
}
\arguments{
\item{pi_s}{the vector of mixture proportions, with first element corresponding to null}

\item{lm}{the results of a likelihood matrix calculation from \code{calc_relative_lik_matrix}
whose first column corresponds to null}
}
\description{
Compute vector of alternative loglikelihoods from a matrix of log-likelihoods and fitted pi
}
