% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_covs.R
\name{compute_covs_allones}
\alias{compute_covs_allones}
\title{Compute an R by R matrix of all 1s}
\usage{
compute_covs_allones(Bhat)
}
\arguments{
\item{Bhat}{an N by R matrix of data}
}
\value{
a list with 1 entry, the R by R matrix of all 1s
export
}
\description{
Compute an R by R matrix of all 1s
}

