% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_covs.R
\name{compute_covs_allzeros}
\alias{compute_covs_allzeros}
\title{Compute an R by R matrix of all 0s}
\usage{
compute_covs_allzeros(data)
}
\arguments{
\item{data}{a mash data object, eg as created by \code{set_mash_data}}
}
\value{
a list with 1 entry, the R by R matrix of all 0s
}
\description{
Compute an R by R matrix of all 0s
}

