% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_covs.R
\name{compute_covs_identity}
\alias{compute_covs_identity}
\title{Compute an identity matrix}
\usage{
compute_covs_identity(Bhat)
}
\arguments{
\item{Bhat}{an N by R matrix of data}
}
\value{
a list with one entry, an R by R identity matrix
export
}
\description{
Compute an identity matrix
}

