% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_covs.R
\name{compute_covs_singletons}
\alias{compute_covs_singletons}
\title{Compute all R singleton matrices corresponding to condition-specific effects}
\usage{
compute_covs_singletons(Bhat)
}
\arguments{
\item{Bhat}{an N by R matrix of data}
}
\value{
a list with R entries, the rth entry being an R by R matrix with all 0s except the (r,r) element is 1
export
}
\description{
Compute all R singleton matrices corresponding to condition-specific effects
}

