% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_cor.R
\name{compute_null_correlation_lower_bound}
\alias{compute_null_correlation_lower_bound}
\title{compute_null_correlation_lower_bound}
\usage{
compute_null_correlation_lower_bound(data)
}
\arguments{
\item{data}{Description of this argument goes here.}
}
\description{
Compute a lower bound on the null correlation matrix from data
}
\details{
Assume $$z^j_r = mu^j_r + e_r$$ where $e_1,e_2$ are joint normal with
variance 1 and some covariance (same as correlation since they are variance 1).
Then $E((z^j_1 - z^j_2)^2) = E(mu^j_1-mu^j_2)^2 + 2(1-cov(e_1,e_2)) > 2(1-cov(e_1,e_2))$.
Thus $$cov(e_1,e_2) > 1- 0.5E((z^j_1 - z^j_2)^2)$$ gives a lower bound on the covariance.
}
