% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{compute_posterior_arrays}
\alias{compute_posterior_arrays}
\title{compute_posterior_arrays}
\usage{
compute_posterior_arrays(data, Ulist)
}
\arguments{
\item{data}{a mash data object, eg as created by \code{set_mash_data}}

\item{Ulist}{list of P prior covariance matrices}
}
\value{
post_means JxPxR array of posterior means

post_vars JxPxR array of posterior (marginal) variances

post_pos JxPxR array of posterior (marginal) probability of being positive

post_neg JxPxR array of posterior (marginal) probability of being negative

post_zero JxPxR array of posterior (marginal) probability of being zero
}

