% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{compute_posterior_weights}
\alias{compute_posterior_weights}
\title{compute posterior probabilities}
\usage{
compute_posterior_weights(prior, lik_mat)
}
\arguments{
\item{prior}{a K vector of prior weights}

\item{lik_mat}{a JxK matrix of likelihoods}
}
\value{
a JxK matrix of posterior probabilities, the jth row contains posteriors for jth effect
}
\details{
computes posterior probabilities that each effect came from each component
}

