% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_covs.R
\name{cov_all_ones}
\alias{cov_all_ones}
\title{Compute an R by R matrix of all 1s}
\usage{
cov_all_ones(data)
}
\arguments{
\item{data}{a mash data object, eg as created by \code{set_mash_data}}
}
\value{
a list with 1 entry, the R by R matrix of all 1s
}
\description{
Compute an R by R matrix of all 1s
}

