% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_covs.R
\name{cov_canonical}
\alias{cov_canonical}
\title{Compute a list of covariance matrices using supplied functions, and put them all together}
\usage{
cov_canonical(data, cov_methods = c("identity", "singletons", "equal_effects",
  "simple_het"), Ulist = NULL)
}
\arguments{
\item{data}{a mash data object, eg as created by \code{set_mash_data}}

\item{cov_methods}{a named list defining the methods (functions and arguments) to be used}

\item{Ulist}{optionally, an existing list of matrices to which the newly computed matrices will be added}
}
\value{
a list of covariance matrices
}
\description{
Compute a list of covariance matrices using supplied functions, and put them all together
}
\details{
Each element of methods may be either a character string which names a pre-defined method (see \code{cov_methods}) or a list with elements fn and args, specifying the function to call, and any additional arguments
}
\examples{
data = set_mash_data(Bhat = cbind(c(1,2),c(3,4)), Shat = cbind(c(1,1),c(1,1)))
 cov_canonical(data,"id")
 cov_canonical(data,"singletons")
 cov_canonical(data,c("id","sing")) # can use partial matching of names
 cov_canonical(data,list("sing","id",eg = list(fn = cov_simple_het, args= list(corr=c(0.1,0.2)))))
}
