% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_covs.R
\name{cov_data_driven}
\alias{cov_data_driven}
\title{Compute a list of data-driven covariance matrices;}
\usage{
cov_data_driven(data, subset = NULL)
}
\arguments{
\item{data}{a mash data object, eg as created by \code{set_mash_data}}
}
\value{
a list of covariance matrices
}
\description{
Compute a list of data-driven covariance matrices;
}
\details{
currently just applies Extreme Deconvolution to a subset of signals; future expansion expected
}
