% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2cov.R
\name{cov_ed}
\alias{cov_ed}
\title{Perform "extreme deconvolution" (Bovy et al) on a subset of the data}
\usage{
cov_ed(data, Ulist_init, subset)
}
\arguments{
\item{data}{a mash data object}

\item{Ulist_init}{a named list of covariance matrices to use to initialize ED; default is to use matrices from  PCs}

\item{subset}{a subset of data to be used when ED is run (set to NULL for all the data)}
}
\description{
Perform "extreme deconvolution" (Bovy et al) on a subset of the data
}
\details{
Runs the extreme deconvolution algorithm from Bovy et al (Annals of Applied Statistics) to estimate data-driven covariance matrices
The default is to initialize the EM algorithm from data2cov_pca with 5 PCs
}
