% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_covs.R
\name{cov_equal_effects}
\alias{cov_equal_effects}
\title{Compute an R by R matrix of all 1s}
\usage{
cov_equal_effects(data)
}
\arguments{
\item{data}{a mash data object, eg as created by \code{mash_set_data}}
}
\value{
a list with 1 entry, the R by R matrix of all 1s
}
\description{
This is an internal (non-exported) function. This help
  page provides additional documentation mainly intended for
  developers and expert users.
}
\keyword{internal}
