% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2cov.R
\name{cov_from_factors}
\alias{cov_from_factors}
\title{produce list of rank-1 covariance matrices corresponding to rows of f}
\usage{
cov_from_factors(f, name)
}
\arguments{
\item{f}{a matrix of factors (each row is a factor)}

\item{name}{a string indicating the name to use}
}
\value{
a list of rank one matrices whose kth element is f[k,]
f[k,]' and named name_k
}
\description{
produce list of rank-1 covariance matrices corresponding to rows of f
}
\keyword{internal}
