% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_covs.R
\name{cov_identity}
\alias{cov_identity}
\title{Compute an identity matrix}
\usage{
cov_identity(data)
}
\arguments{
\item{data}{a mash data object, eg as created by \code{set_mash_data}}
}
\value{
the R by R identity matrix
}
\description{
Compute an identity matrix
}

