% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_covs.R
\name{cov_methods}
\alias{cov_methods}
\title{A function that maps names of covariance matrices to covariance computation functions}
\usage{
cov_methods()
}
\value{
a named list of defaults for covariance function calculation
}
\description{
A function that maps names of covariance matrices to covariance computation functions
}
