% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_covs.R
\name{cov_simple_het}
\alias{cov_simple_het}
\title{For each element of corr compute a matrix with diagonal element 1 and off-diagonal element corr}
\usage{
cov_simple_het(data, corr)
}
\arguments{
\item{data}{a mash data object, eg as created by \code{set_mash_data}}
}
\value{
a list of matrices
}
\description{
For each element of corr compute a matrix with diagonal element 1 and off-diagonal element corr
}

