% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_covs.R
\name{cov_singletons}
\alias{cov_singletons}
\title{Compute all R singleton matrices corresponding to condition-specific effects}
\usage{
cov_singletons(data)
}
\arguments{
\item{data}{a mash data object, eg as created by \code{set_mash_data}}
}
\value{
a list with R entries, the rth entry being an R by R matrix with all 0s except the (r,r) element is 1
}
\description{
Compute all R singleton matrices corresponding to condition-specific effects
}

