% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed.R
\name{ed_wrapper}
\alias{ed_wrapper}
\title{Fit extreme deconvolution to mash data}
\usage{
ed_wrapper(data, Ulist_init, subset = NULL, ...)
}
\arguments{
\item{data}{mash data object}

\item{Ulist_init}{a list of covariance matrices to initialize to}

\item{subset}{the indices of the observations to be used (defaults to all of them)}

\item{...}{arguments to be passed to \code{extreme_deconvolution} function, such as \code{tol}, \code{maxiter}. Other options see \code{?ExtremeDeconvolution::extreme_deconvolution}}
}
\value{
the fitted mixture: a list of mixture proportions and covariance matrices
}
\description{
Fit extreme deconvolution to mash data
}
\details{
This is a wrapper to ExtremeDeconvolution::extreme_deconvolution
It fixes the projection to be the identity, and the means to be 0
}
