% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_cor.R
\name{estimate_null_correlation}
\alias{estimate_null_correlation}
\title{Estimate null correlations}
\usage{
estimate_null_correlation(data, z_thresh = 2, apply_lower_bound = TRUE)
}
\arguments{
\item{data}{a mash data object, eg as created by \code{set_mash_data}}

\item{z_thresh}{the z score threshold below which to call an effect null}

\item{apply_lower_bound}{indicates whether to also apply a lower bound for the
correlations which can be computed based on all data. See \code{compute_null_correlation_lower_bound}}
}
\description{
Estimates a null correlation matrix from data
}
\details{
Returns the empirical correlation matrix of the effects that are "null" based on simple z score threshold
}
