% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_cor.R
\name{estimate_null_correlation_adhoc}
\alias{estimate_null_correlation_adhoc}
\title{Estimate null correlations (ad hoc)}
\usage{
estimate_null_correlation_adhoc(data, z_thresh = 2)
}
\arguments{
\item{data}{a mash data object, eg as created by \code{mash_set_data}}

\item{z_thresh}{the z score threshold below which to call an effect null}
}
\description{
Estimates a null correlation matrix from data using simple z score threshold
}
\details{
Returns the empirical correlation matrix of the effects that are "null" based on simple z score threshold
}
