% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreme_deconvolution.R
\name{extreme_deconvolution}
\alias{extreme_deconvolution}
\title{Density estimation using Gaussian mixtures in the presence of
  noisy, heterogeneous and incomplete data}
\usage{
extreme_deconvolution(ydata, ycovar, xamp, xmean, xcovar,
  projection = NULL, weight = NULL, fixamp = NULL, fixmean = NULL,
  fixcovar = NULL, tol = 1e-06, maxiter = 1e+09, w = 0,
  logfile = NULL, splitnmerge = 0, maxsnm = FALSE,
  likeonly = FALSE, logweight = FALSE)
}
\arguments{
\item{ydata}{[ndata,dy] matrix of observed quantities.}

\item{ycovar}{[ndata,dy] / [ndata,dy,dy] / [dy,dy,ndata] matrix,
list or 3D array of observational error covariances (if [ndata,dy]
then the error correlations are assumed to vanish).}

\item{xamp}{[ngauss] array of initial amplitudes (*not* [1,ngauss]).}

\item{xmean}{[ngauss,dx] matrix of initial means.}

\item{xcovar}{[ngauss,dx,dx] list of matrices of initial covariances.}

\item{projection}{[ndata,dy,dx] list of projection matrices.}

\item{weight}{[ndata] array of weights to be applied to the data points.}

\item{fixamp}{None, True/False, or list of bools.}

\item{fixmean}{None, True/False, or list of bools.}

\item{fixcovar}{None, True/False, or list of bools.}

\item{tol}{(double) Tolerance for convergence.}

\item{maxiter}{Maximum number of iterations to perform.}

\item{w}{(double) covariance regularization parameter (of the
conjugate prior).}

\item{logfile}{Basename for several logfiles (_c.log has output
from the c-routine; _loglike.log has the log likelihood path of all
the accepted routes, i.e. only parts which increase the likelihood
are included, during splitnmerge).}

\item{splitnmerge}{(int, default=0) depth to go down the
splitnmerge path.}

\item{likeonly}{(Bool) Only compute the total log likelihood of the
data.}

\item{logweight}{(bool) if True, weight is actually
log(weight).}

\item{(Bool)}{Use the maximum number of split 'n' merge steps,
K*(K-1)*(K-2)/2.}
}
\value{
A list object with the following elements:
\item{avgloglikedata}{avgloglikedata after convergence}
\item{xamp}{updated xamp}
\item{xmean}{updated xmean}
\item{xcovar}{updated xcovar}
}
\description{
We present a general algorithm to infer a
  d-dimensional distribution function given a set of heterogeneous,
  noisy observations or samples. This algorithm reconstructs the
  error-deconvolved or "underlying" distribution function common to
  all samples, even when the individual samples have unique error and
  missing-data properties. The underlying distribution is modeled as
  a mixture of Gaussians, which is completely general. Model
  parameters are chosen to optimize a justified, scalar objective
  function: the logarithm of the probability of the data under the
  error-convolved model, where the error convolution is different for
  each data point. Optimization is performed by an Expectation
  Maximization (EM) algorithm, extended by a regularization technique
  and "split-and-merge" procedure. These extensions mitigate problems
  with singularities and local maxima, which are often encountered
  when using the EM algorithm to estimate Gaussian density mixtures.
}
\references{
Inferring complete distribution functions from noisy, heterogeneous
and incomplete observations Jo Bovy, David W. Hogg, & Sam
T. Roweis, Submitted to AOAS (2009) [arXiv/0905.2979]
}
\author{
Jo Bovy, David W. Hogg & Sam T. Roweis.
}
