% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2cov.R
\name{factors2cov}
\alias{factors2cov}
\title{produce list of rank 1 covariance matrices corresponding to rows of f}
\usage{
factors2cov(f, name)
}
\description{
produce list of rank 1 covariance matrices corresponding to rows of f
}

