% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_data.R
\name{get_cov}
\alias{get_cov}
\title{return the covariance matrix for jth data point from mash data object}
\usage{
get_cov(data, j)
}
\description{
return the covariance matrix for jth data point from mash data object
}
