% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_expanded_cov}
\alias{get_expanded_cov}
\title{Get expanded list of covariance matrices in m, expanded by grid}
\usage{
get_expanded_cov(m)
}
\arguments{
\item{m}{a mash object}
}
\value{
a list of covariance matrices
This takes the covariance matrix in m and multiplies them by the grid in m
If a pointmass is included in m then it adds a null component
}
\description{
Get expanded list of covariance matrices in m, expanded by grid
}

