% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_lfsr}
\alias{get_lfsr}
\title{Extract lfdr from mash object}
\usage{
get_lfsr(m)
}
\arguments{
\item{m}{a mash object}
}
\value{
a matrix of lfsr values, with (j,r)th entry corresponding to the lfsr for effect j in condition r
}
\description{
Extract lfdr from mash object
}
