% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_mash_fitted_g}
\alias{get_mash_fitted_g}
\title{Return the fitted g from a mash object}
\usage{
get_mash_fitted_g(m)
}
\arguments{
\item{m}{a mash object, as returned by \code{mash}}
}
\description{
Return the fitted g from a mash object
}
