% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_functions.R
\name{get_mixprob}
\alias{get_mixprob}
\title{Return the mixture proportions in g}
\usage{
get_mixprob(g)
}
\arguments{
\item{g}{a mixture of multivariate normals, as created for example by \ode{initialize_g}}
}
\description{
Return the mixture proportions in g
}

