% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_n_significant_conditions}
\alias{get_n_significant_conditions}
\title{Count number of conditions each effect is significant in}
\usage{
get_n_significant_conditions(
  m,
  thresh = 0.05,
  conditions = NULL,
  sig_fn = get_lfsr
)
}
\arguments{
\item{m}{the mash result (from joint or 1by1 analysis)}

\item{thresh}{indicates the threshold below which to call signals significant}

\item{conditions}{which conditions to include in check (default to all)}

\item{sig_fn}{the significance function used to extract significance from mash object; eg could be ashr::get_lfsr or ashr::get_lfdr}
}
\value{
a vector containing the number of significant conditions
}
\description{
Count number of conditions each effect is significant in
}
