% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_posterior_matrices}
\alias{get_posterior_matrices}
\title{Extract posterior matrices from mash object}
\usage{
get_posterior_matrices(m)
}
\arguments{
\item{m}{a mash object}
}
\value{
a list of posterior matrices computed
}
\description{
Extract posterior matrices from mash object
}
