% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_samples}
\alias{get_samples}
\title{Return samples from a mash object}
\usage{
get_samples(m)
}
\arguments{
\item{m}{The mash fit.}
}
\description{
Return samples from a mash object
}
\examples{
simdata = simple_sims(50,5,1)
data = mash_set_data(simdata$Bhat, simdata$Shat)
m = mash(data, cov_canonical(data), posterior_samples=5, algorithm='R')
get_samples(m)
}
