% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_Shat}
\alias{get_Shat}
\title{Return matrix of Shat values from mash object}
\usage{
get_Shat(m, subset.eff = NULL, subset.cond = NULL)
}
\arguments{
\item{m}{a mash object}

\item{subset.eff}{specifies the subset of effects (rows) to include}

\item{subset.cond}{specifies the subset of conditions (columns) to include}
}
\description{
Return matrix of Shat values from mash object
}
