% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_mash.R
\name{get_significant_results}
\alias{get_significant_results}
\title{Find effects that have lfsr < thresh in at least one condition}
\usage{
get_significant_results(m, thresh = 0.05)
}
\arguments{
\item{m}{the mash result (from joint or 1by1 analysis)}

\item{thresh}{indicates the threshold below which to set signals}
}
\value{
a vector containing the indices of the significant effects
}
\description{
Find effects that have lfsr < thresh in at least one condition
}
