% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_functions.R
\name{initialize_g}
\alias{initialize_g}
\title{Create a}
\usage{
initialize_g(data, Umethods = c("all_zero", "id", "singletons", "bmalite",
  "all_one"), grid)
}
\arguments{
\item{data}{a mash data object, e.g. as created by \code{set_mash_data}}

\item{Umethods}{a vector of strings that indicates which methods to use to create covariance matrices}

\item{grid}{a vector of scalars that are to be used to scale the covariance matrices}
}
\description{
Create a
}

