% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_mash.R
\name{initialize_pi}
\alias{initialize_pi}
\title{Initialize mixture proportions - currently by making them all equal}
\usage{
initialize_pi(K)
}
\arguments{
\item{K}{the number of components}
}
\value{
a vector of length K whose elements are positive and sums to 1
}
\description{
Initialize mixture proportions - currently by making them all equal
}
