% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods_common_cov.R
\name{is_common_cov}
\alias{is_common_cov}
\title{checks if all rows of Shat are the same - if so covariances are equal}
\usage{
is_common_cov(data)
}
\description{
checks if all rows of Shat are the same - if so covariances are equal
}
