% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods_common_cov.R
\name{is_common_cov_Shat_alpha}
\alias{is_common_cov_Shat_alpha}
\title{Check that all covariates are equal (Shat_alpha).}
\usage{
is_common_cov_Shat_alpha(data)
}
\arguments{
\item{data}{A mash data object.}
}
\description{
checks if all rows of Shat_alpha are the same - if so
    covariances are equal
}
