% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mash.R
\name{list_cov}
\alias{list_cov}
\title{List names of covariance matrices in m}
\usage{
list_cov(m)
}
\arguments{
\item{m}{a mash object}
}
\value{
names of covariance matrices in m
}
\description{
List names of covariance matrices in m
}
