% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cov.R
\name{make_names}
\alias{make_names}
\title{Create names for covariance matrices}
\usage{
make_names(names, suffixes)
}
\arguments{
\item{names}{a string}

\item{suffixes}{Adds _suffixes to names for each element of suffixes}
}
\description{
Create names for covariance matrices
}
