% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cov.R
\name{mash_add_cov_ed}
\alias{mash_add_cov_ed}
\title{Add covariance matrices to m from "extreme deconvolution" (Bovy et al)}
\usage{
mash_add_cov_ed(m, Ulist_init, subset)
}
\arguments{
\item{m}{a mash object}

\item{Ulist_init}{a named list of covariance matrices to use to initialize ED}

\item{subset}{a subset of data to be used when ED is run}
}
\description{
Add covariance matrices to m from "extreme deconvolution" (Bovy et al)
}
\details{

}
