% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cov.R
\name{mash_add_cov_equal_effects}
\alias{mash_add_cov_equal_effects}
\title{Add a covariance matrix of equal effects to m}
\usage{
mash_add_cov_equal_effects(m)
}
\arguments{
\item{m}{a mash object}
}
\description{
Add a covariance matrix of equal effects to m
}

