% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cov.R
\name{mash_add_cov_identity}
\alias{mash_add_cov_identity}
\title{Add identify covariance matrix to m}
\usage{
mash_add_cov_identity(m)
}
\arguments{
\item{m}{a mash object}
}
\description{
Add identify covariance matrix to m
}
