% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cov.R
\name{mash_add_cov_list}
\alias{mash_add_cov_list}
\title{Add a list of covariance matrices to m}
\usage{
mash_add_cov_list(m, Ulist)
}
\arguments{
\item{m}{a mash object}

\item{Ulist}{a named list of covariance matrices to add to m}
}
\description{
Add a list of covariance matrices to m
}
\details{
This allows the user to provide their own list of covariance matrices if they desire,
in addition to built-in covariance options
}

