% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cov.R
\name{mash_add_cov_pca}
\alias{mash_add_cov_pca}
\title{Add covariance matrices by doing PCA on data in mash object}
\usage{
mash_add_cov_pca(m, k = 5, subset = NULL)
}
\arguments{
\item{m}{a mash object}

\item{k}{number of PCs to use}

\item{subset}{subset of data points to use in PCA (default of NULL causes top hits in m to be used)}
}
\description{
Add covariance matrices by doing PCA on data in mash object
}
