% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cov.R
\name{mash_add_cov_r1}
\alias{mash_add_cov_r1}
\title{Add rank 1 covariance matrices to m}
\usage{
mash_add_cov_r1(m, f)
}
\arguments{
\item{m}{a mash object}

\item{f}{a matrix, whose rows are each of length R; if f is a vector, co-erced to a matrix with one row}
}
\description{
Add rank 1 covariance matrices to m
}
\details{
For each row x of matrix f, this adds the rank one matrix xx' as a covariance matrix to m
}
