% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cov.R
\name{mash_add_cov_simple_het}
\alias{mash_add_cov_simple_het}
\title{Add covariance matrices allowing for heterogenous (positively correlated) effects}
\usage{
mash_add_cov_simple_het(m)
}
\arguments{
\item{m}{a mash object}
}
\description{
Add covariance matrices allowing for heterogenous (positively correlated) effects
}
\details{
These covariance matrices assume all pairs of conditions are equally correlated.
}

