% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cov.R
\name{mash_add_cov_singletons}
\alias{mash_add_cov_singletons}
\title{Add singleton covariance matrices to m}
\usage{
mash_add_cov_singletons(m)
}
\arguments{
\item{m}{a mash object}
}
\description{
Add singleton covariance matrices to m
}
