% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mash.R
\name{mash_add_grid}
\alias{mash_add_grid}
\title{Add a grid of scaling factors to a mash object}
\usage{
mash_add_grid(m, grid)
}
\arguments{
\item{m}{the mash object}

\item{grid}{a vector of numeric scaling factors}
}
\description{
Add a grid of scaling factors to a mash object
}
\details{
These scaling factors are used to scale all
covariance matrices in m when fitting the model.
They correspond to the omega values in Urbut et al.
}

