% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mash.R
\name{mash_add_strong_signals}
\alias{mash_add_strong_signals}
\title{Add a list of strong signals (to be used in data-drive covariance matrices)}
\usage{
mash_add_strong_signals(m, thresh = 0.05)
}
\arguments{
\item{m}{the mash object}

\item{thresh}{indicates the threshold below which to set signals}
}
\description{
Add a list of strong signals (to be used in data-drive covariance matrices)
}
\details{
Adds the top univariate signals (lfsr<0.05)
}
