% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mash.R
\name{mash_compute_loglik}
\alias{mash_compute_loglik}
\title{Compute loglikelihood for fitted mash object}
\usage{
mash_compute_loglik(m, Bhat = NULL, Shat = NULL)
}
\arguments{
\item{m}{a mash object}

\item{Bhat}{a matrix of Bhat values}

\item{Shat}{a matrix of corresponding Shat values}
}
\description{
Compute loglikelihood for fitted mash object
}
\details{
If no data are supplied, then computes log-likelihood on data in m (ie training data).
If data (Bhat and Shat) are supplied then computes log-likelihood under the model fit in m.
The latter might be useful for computing log-likelihood on a "test" set for example.
}
