% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mash.R
\name{mash_compute_posterior}
\alias{mash_compute_posterior}
\title{Compute posterior quantities for mash object}
\usage{
mash_compute_posterior(m)
}
\arguments{
\item{m}{a mash object}
}
\description{
Compute posterior quantities for mash object
}
\details{
Computes posterior weights and posterior matrices (eg posterior mean etc)
Uses data (Bhat and Shat) added to m and the fitted g obtained by \code{mash_fit_g}
}

