% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mash.R
\name{mash_compute_vloglik}
\alias{mash_compute_vloglik}
\title{Compute vector of loglikelihood for fitted mash object on new data}
\usage{
mash_compute_vloglik(g, data)
}
\arguments{
\item{g}{a mash object or the fitted_g from a mash object}

\item{data}{a set of data on which to compute the loglikelihood}
}
\value{
the vector of log-likelihoods for each data point computed using g
}
\description{
Compute vector of loglikelihood for fitted mash object on new data
}
