% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mash.R
\name{mash_fit_g}
\alias{mash_fit_g}
\title{Fit the Empirical Bayes model for a mash object}
\usage{
mash_fit_g(m, prior = "nullbiased", optmethod = c("mixIP"))
}
\arguments{
\item{m}{a mash object}
}
\description{
Fit the Empirical Bayes model for a mash object
}
\details{
Computes the likelihood matrix and fits the
mixture proportions for a mixture of multivariate normals.
The mash object contains the data, and the covariance matrices and grid to use
See \code{mash_add_cov} and \code{mash_add_grid}
}

