% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mash.R
\name{mash_init}
\alias{mash_init}
\title{Initialize a mash object (actually an environment)}
\usage{
mash_init(Bhat, Shat)
}
\arguments{
\item{R}{the number of conditions to be used}
}
\value{
a mash object
}
\description{
Initialize a mash object (actually an environment)
}

