% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_mash.R
\name{mash_new}
\alias{mash_new}
\title{Apply mash method to data}
\usage{
mash_new(data, Ulist, gridmult = sqrt(2), grid = NULL, normalizeU = TRUE,
  usepointmass = TRUE, prior = c("uniform", "nullbiased"),
  optmethod = c("mixIP"))
}
\arguments{
\item{data}{a mash data object containing the Bhat matrix and standard errors}

\item{Ulist}{a list of covariance matrices to use}

\item{gridmult}{scalar indicating factor by which adjacent grid values should differ; close to 1 for fine grid}

\item{grid}{vector of grid values to use (scaling factors omega in paper)}

\item{prior}{indicates what penalty to use on the likelihood, if any}

\item{optmethod}{name of optimization method to use}
}
\description{
Apply mash method to data
}
