% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mash.R
\name{mash_normalize_Ulist}
\alias{mash_normalize_Ulist}
\title{Normalize the covariances in mash object}
\usage{
mash_normalize_Ulist(m)
}
\arguments{
\item{m}{the mash object}
}
\description{
Normalize the covariances in mash object
}
\details{
Normalizes the covariance matrices to have max diagonal element 1
}
