% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mash_run_1by1.R
\name{mash_run_1by1}
\alias{mash_run_1by1}
\title{Perform condition-by-condition analyses}
\usage{
mash_run_1by1(m, force = FALSE)
}
\arguments{
\item{m}{a mash object}
}
\value{
Adds to m the results (posterior_matrices) from the ash runs
}
\details{
Performs "condition-by-condition" analysis
by running \code{ash} from package \code{ashr} on data from each condition, one at a time.
May be a useful first step to identify top hits in each condition before a mash analysis.
}

