% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_mash.R
\name{mash_run_1by1_new}
\alias{mash_run_1by1_new}
\title{Perform condition-by-condition analyses}
\usage{
mash_run_1by1_new(data)
}
\arguments{
\item{Bhat}{an n by R matrix of observations (n units in R conditions)}

\item{Shat}{an n by R matrix of standard errors (n units in R conditions)}
}
\value{
posterior_matrices from the ash runs
}
\details{
Performs simple "condition-by-condition" analysis
by running \code{ash} from package \code{ashr} on data from each condition, one at a time.
May be a useful first step to identify top hits in each condition before a mash analysis.
}
