% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_data.R
\name{mash_set_data_contrast}
\alias{mash_set_data_contrast}
\title{Create a data object for mash contrast analysis}
\usage{
mash_set_data_contrast(mashdata, L)
}
\arguments{
\item{mashdata}{a mash data object containing the Bhat matrix,
standard errors, V; created using \code{mash_set_data}}

\item{L}{the contrast matrix}
}
\value{
a data object after the contrast transfermation
}
\description{
This is an internal (non-exported) function. This help
  page provides additional documentation mainly intended for
  developers and expert users.
}
\keyword{internal}
