% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mash.R
\name{mash_status}
\alias{mash_status}
\title{Summarize status of a mash analysis}
\usage{
mash_status(m = NULL)
}
\arguments{
\item{m}{a mash object}
}
\description{
Summarize status of a mash analysis
}
\details{
Provides a summary of what has been done so far in a mash object; may provide
suggestions for next step
}
