% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R, R/posterior.R
\name{mashr-internal}
\alias{mashr-internal}
\alias{posterior_cov}
\alias{posterior_mean}
\alias{posterior_mean_matrix}
\alias{compute_posterior_weights}
\title{Internal Package Functions}
\usage{
posterior_cov(Vinv, U)

posterior_mean(bhat, Vinv, U1)

posterior_mean_matrix(Bhat, Vinv, U1)

compute_posterior_weights(pi, lik_mat)
}
\arguments{
\item{Vinv}{R x R inverse covariance matrix for the likelihood}

\item{U}{R x R prior covariance matrix}

\item{bhat}{R vector of observations}

\item{U1}{R x R posterior covariance matrix, computed using posterior_cov}

\item{Bhat}{J by R matrix of observations}

\item{pi}{a K vector of mixture proportions}

\item{lik_mat}{a JxK matrix of likelihoods}
}
\value{
R x R posterior covariance matrix

R vector of posterior mean

R vector of posterior mean

a JxK matrix of posterior probabilities, the jth row contains posteriors for jth effect
}
\description{
This help page provides additional documentation of
internal (non-exported) functions, mainly for developers and expert
users.

If bhat is N(b,V) and b is N(0,U) then b|bhat N(mu1,U1). This function returns U1.

If bhat is N(b,V) and b is N(0,U) then b|bhat N(mu1,U1). This function returns mu1.

Computes posterior mean under multivariate normal model for each row of matrix Bhat.
Note that if bhat is N_R(b,V) and b is N_R(0,U) then b|bhat N_R(mu1,U1).
This function returns a matrix with jth row equal to mu1(bhat) for bhat= Bhat[j,].

computes posterior probabilities that each effect came from each component
}
\keyword{internal}
