% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_covs.R
\name{multiply_list}
\alias{multiply_list}
\title{multiply each element of a list by scalar
(In our application each element of the list is a matrix)}
\usage{
multiply_list(Ulist, x)
}
\description{
multiply each element of a list by scalar
(In our application each element of the list is a matrix)
}
