% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_covs.R
\name{normalize_cov}
\alias{normalize_cov}
\title{normalize a covariance matrix soits maximum diagonal element is 1
Divides each element of the matrix by its maximum diagonal element (provided that it is non-zero)}
\usage{
normalize_cov(U)
}
\description{
normalize a covariance matrix soits maximum diagonal element is 1
Divides each element of the matrix by its maximum diagonal element (provided that it is non-zero)
}

