% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_covs.R
\name{normalize_Ulist}
\alias{normalize_Ulist}
\title{Normalize a list of covariance matrices by applying normalize_cov
to each element.}
\usage{
normalize_Ulist(Ulist)
}
\description{
Normalize a list of covariance matrices by applying normalize_cov
to each element.
}
