% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt.R
\name{optimize_pi}
\alias{optimize_pi}
\title{Estimate mixture weights by maximum (penalized) likelihood}
\usage{
optimize_pi(
  matrix_lik,
  pi_init = NULL,
  prior = NULL,
  optmethod = c("mixSQP", "mixIP", "mixEM", "cxxMixSquarem"),
  control = list()
)
}
\arguments{
\item{matrix_lik}{a matrix of likelihoods, where the (i,k)th entry
is the probability of observation i given it came from component k
of g}

\item{pi_init}{numeric vector specifying value from which to
initialize optimization}

\item{prior}{numeric vector specifying prior to use in the
penalized likelihood}

\item{optmethod}{a string, giving name of optimization function to
use}

\item{control}{a list of parameters to be passed to optmethod}
}
\value{
numeric vector specifying the optimal mixture weights
}
\description{
This is an internal (non-exported) function. This help
  page provides additional documentation mainly intended for
  developers and expert users.
}
\keyword{internal}
