% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_meta}
\alias{plot_meta}
\title{Plot metaplot for an effect based on posterior from mash}
\usage{
plot_meta(m, i, xlab = "Effect size", ylab = "Condition", ...)
}
\arguments{
\item{m}{the result of a mash fit}

\item{i}{index of the effect to plot}
}
\description{
Plot metaplot for an effect based on posterior from mash
}
