% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{posterior_mean}
\alias{posterior_mean}
\title{posterior_mean}
\usage{
posterior_mean(bhat, Vinv, U1)
}
\arguments{
\item{bhat}{R vector of observation}

\item{Vinv}{R x R inverse covariance matrix for the likelihood}

\item{U1}{R x R posterior covariance matrix, computed using posterior_cov}
}
\value{
R vector of posterior mean
}
\description{
This is an internal (non-exported) function. This help
  page provides additional documentation mainly intended for
  developers and expert users.
}
\details{
If bhat is N(b,V) and b is N(0,U) then b|bhat
  N(mu1,U1). This function returns mu1.
}
\keyword{internal}
