% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_functions.R
\name{print_biggest_comp}
\alias{print_biggest_comp}
\title{Print out the components of g with largest weight (those exceeding thresh)}
\usage{
print_biggest_comp(g, thresh = 0.01)
}
\arguments{
\item{g}{a mixture of multivariate normals, e.g. as created by \code{add_to_g}}

\item{thresh}{the threshold on mixture weight; only components exceeding weight are output}
}
\description{
Print out the components of g with largest weight (those exceeding thresh)
}

