% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_covs.R
\name{scale_cov}
\alias{scale_cov}
\title{Scale each covariance matrix in list Ulist by a scalar in vector grid}
\usage{
scale_cov(Ulist, grid)
}
\arguments{
\item{Ulist}{a list of matrices}

\item{grid}{a vector of scaling factors}
}
\value{
a list with length length(Ulist)*length(grid), with values grid[i]*Ulist[[j]]
}
\description{
Scale each covariance matrix in list Ulist by a scalar in vector grid
}

