% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_covs.R
\name{scale_Ulist}
\alias{scale_Ulist}
\title{Scale each matrix in list Ulist by a scalar in vector grid}
\usage{
scale_Ulist(Ulist, grid)
}
\arguments{
\item{Ulist}{a list of matrices}

\item{grid}{a vector of scaling factors}
}
\value{
a list with length length(Ulist) \times length(grid), with values grid[i]\times Ulist[[j]]
}
\description{
Scale each matrix in list Ulist by a scalar in vector grid
}

