% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_data.R
\name{set_mash_data}
\alias{set_mash_data}
\title{Deprecated function call of `mash_set_data`.}
\usage{
set_mash_data(Bhat, Shat = NULL, alpha = 0, df = Inf, pval = NULL,
  V = diag(ncol(Bhat)))
}
\description{
Deprecated function call of `mash_set_data`.
}
