% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_mash.R
\name{set_prior}
\alias{set_prior}
\title{sets prior to be a vector of length K depending on character string
prior can be "nullbiased" or "uniform"}
\usage{
set_prior(K, prior)
}
\description{
sets prior to be a vector of length K depending on character string
prior can be "nullbiased" or "uniform"
}
