% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{simple_sims2}
\alias{simple_sims2}
\title{Create some more simple simulated data for testing purposes}
\usage{
simple_sims2(nsamp = 100, err_sd = 0.01)
}
\arguments{
\item{nsamp}{number of samples of each type}

\item{err_sd}{the standard deviation of the errors}
}
\description{
Create some more simple simulated data for testing purposes
}
\details{
The simulation consists of five conditions with two types
of effecc those present (and identical) in first two conditions and
those present (and identical) in last three conditions
}
\examples{
simple_sims2(100, 5)

}
