% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed.R
\name{teem_wrapper}
\alias{teem_wrapper}
\title{Fit extreme deconvolution to mash data using TEEM method
developed by Y. Yang and M Stephens}
\usage{
teem_wrapper(
  data,
  Ulist_init,
  subset = NULL,
  w_init = NULL,
  maxiter = 5000,
  converge_tol = 1e-07,
  eigen_tol = 1e-07,
  verbose = FALSE
)
}
\arguments{
\item{data}{mash data object}

\item{Ulist_init}{a list of covariance matrices to initialize to}

\item{subset}{the indices of the observations to be used (defaults
to all of them)}
}
\value{
the fitted mixture: a list of mixture proportions and
covariance matrices
}
\description{
This is an internal (non-exported) function. This help
  page provides additional documentation mainly intended for
  developers and expert users.
}
\keyword{internal}
