% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_udi.R
\name{udi_model_matrix}
\alias{udi_model_matrix}
\title{Create a matrix whose rows contain all possible
combinations of the U,D,I models that are allowed.}
\usage{
udi_model_matrix(R)
}
\arguments{
\item{R}{the number of conditions}
}
\value{
a matrix that is Nmodel by R with each row containing a
vector of "U", "D" and "I" characters corresponding to a valide
model. Constraint is that there is at least one "D" in each row
}
\description{
This is an internal (non-exported) function. This help
  page provides additional documentation mainly intended for
  developers and expert users.
}
\keyword{internal}
