% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{calc_lik_matrix}
\alias{calc_lik_matrix}
\title{calc_lik_matrix}
\usage{
calc_lik_matrix(data, Ulist, log = FALSE)
}
\arguments{
\item{data}{a mash data object, eg as created by \code{set_mash_data}}

\item{Ulist}{list of prior covariance matrices}

\item{log}{if true computes log-likelihood}
}
\value{
J x P vector of multivariate normal likelihoods, p(bhat_j | Ulist[p], V_j)
}
\description{
computes matrix of likelihoods for each of J rows of Bhat for each of P prior covariances
}
