% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{calc_relative_lik_matrix}
\alias{calc_relative_lik_matrix}
\title{Calculate matrix of relative likelihoods (likelihoods, normalized to avoid numeric issues)}
\usage{
calc_relative_lik_matrix(data, Ulist)
}
\arguments{
\item{data}{a mash data object, eg as created by \code{set_mash_data}}

\item{Ulist}{list of prior covariance matrices}
}
\value{
J x P matrix of likelihoods, p(bhat[j] | Ulist[p], V), but normalized so that the max in each row is 1
}
\details{
computes matrix of relative likelihoods for each of J rows of Bhat for each of P prior covariances
}
