% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_covs.R
\name{compute_covs_identity}
\alias{compute_covs_identity}
\title{Compute an identity matrix}
\usage{
compute_covs_identity(data)
}
\arguments{
\item{data}{a mash data object, eg as created by \code{set_mash_data}}
}
\value{
a list with one entry, an R by R identity matrix
}
\description{
Compute an identity matrix
}

