% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{compute_posterior_arrays}
\alias{compute_posterior_arrays}
\title{compute_posterior_arrays}
\usage{
compute_posterior_arrays(data, Ulist)
}
\arguments{
\item{Ulist}{list of P prior covariance matrices}

\item{Bhat}{J by R matrix of estimates}

\item{Shat}{J x R matrix of standard errors}
}
\value{
post_means JxPxR array of posterior means

post_vars JxPxR array of posterior (marginal) variances

post_pos JxPxR array of posterior (marginal) probability of being positive

post_neg JxPxR array of posterior (marginal) probability of being negative

post_zero JxPxR array of posterior (marginal) probability of being zero
}

