% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{compute_posterior_matrices}
\alias{compute_posterior_matrices}
\title{Compute posterior matrices.}
\usage{
compute_posterior_matrices(data, Ulist, posterior_weights,
  algorithm.version = c("Rcpp", "R"), A = NULL,
  output_posterior_cov = FALSE)
}
\arguments{
\item{data}{A \code{mash} data object; e.g., created by
\code{\link{mash_set_data}} or \code{\link{mash_set_data_contrast}}.}

\item{Ulist}{List containing the prior covariance matrices.}

\item{posterior_weights}{Vector containing the posterior
probability of each mixture component in Ulist for the data}

\item{algorithm.version}{Indicates whether to use R or Rcpp version}

\item{A}{the linear transformation matrix, K x R matrix. This is used to compute the posterior for Ab.}

\item{output_posterior_cov}{whether or not to output posterior covariance matrices for all effects}
}
\value{
The return value is a list containing the following
   components:

   \item{PosteriorMean}{J x K matrix of posterior means.}

   \item{PosteriorSD}{J x K matrix of posterior (marginal) standard
   deviations.}

   \item{NegativeProb}{J x K matrix of posterior (marginal)
    probability of being negative.}

   \item{ZeroProb}{J x K matrix of posterior (marginal) probability
    of being zero.}

   \item{lfsr}{J x K matrix of local false sign rates.}

   \item{PosteriorCov}{K x K x J array of posterior covariance matrices, if the \code{output_posterior_cov = TRUE}.}
}
\description{
More detailed description of function goes here.
}
