% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{compute_posterior_matrices}
\alias{compute_posterior_matrices}
\title{Compute posterior matrices.}
\usage{
compute_posterior_matrices(data, Ulist, posterior_weights,
  algorithm.version = c("Rcpp", "R"))
}
\arguments{
\item{data}{A \code{mash} data object; e.g., created by
\code{\link{set_mash_data}}.}

\item{Ulist}{List containing the prior covariance matrices.}

\item{posterior_weights}{Vector containing the posterior
probability of each mixture component in Ulist for the data}

\item{algorithm.version}{Explain what this argument is for.}
}
\value{
The return value is a list containing the following
   components:

   \item{PosteriorMean}{J x R matrix of posterior means.}

   \item{PosteriorSD}{J x R matrix of posterior (marginal) standard
   deviations.}

   \item{NegativeProb}{J x R matrix of posterior (marginal)
    probability of being negative.}

   \item{ZeroProb}{J x R matrix of posterior (marginal) probability
    of being zero.}

   \item{lfsr}{J x R matrix of local false sign rates.}
}
\description{
More detailed description of function goes here.
}
