% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{compute_posterior_matrices}
\alias{compute_posterior_matrices}
\title{Compute posterior matrices}
\usage{
compute_posterior_matrices(data, Ulist, posterior_weights)
}
\arguments{
\item{data}{a mash data object, eg as created by \code{set_mash_data}}

\item{Ulist}{a list of covariance matrices for each mixture component}

\item{posterior_weights}{the posterior probabilities of each mixture component in Ulist for the data}
}
\value{
post_mean JxR matrix of posterior means

post_sd JxR matrix of posterior (marginal) standard deviations

post_pos JxR matrix of posterior (marginal) probability of being positive

post_neg JxR matrix of posterior (marginal) probability of being negative

post_zero JxR matrix of posterior (marginal) probability of being zero
}
