% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{compute_posterior_weights}
\alias{compute_posterior_weights}
\title{Compute posterior probabilities that each effect came from
  each component}
\usage{
compute_posterior_weights(pi, lik_mat)
}
\arguments{
\item{pi}{a K vector of mixture proportions}

\item{lik_mat}{a JxK matrix of likelihoods}
}
\value{
a JxK matrix of posterior probabilities, the jth row
  contains posteriors for jth effect
}
\description{
This is an internal (non-exported) function. This help
  page provides additional documentation mainly intended for
  developers and expert users.
}
\keyword{internal}
