% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_covs.R
\name{compute_Ulist}
\alias{compute_Ulist}
\title{Compute a list of covariance matrices using supplied functions, and put them all together}
\usage{
compute_Ulist(data, fns = c(compute_covs_identity, compute_covs_singletons,
  compute_covs_allones), Ulist = NULL)
}
\arguments{
\item{data}{a mash data object, eg as created by \code{set_mash_data}}

\item{fns}{a vector of functions to be applied to Bhat to compute covariance matrices}

\item{Ulist}{optionally, an existing list of matrices to which the newly computed matrices will be added}
}
\value{
a list of covariance matrices
}
\description{
Compute a list of covariance matrices using supplied functions, and put them all together
}

