% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_covs.R
\name{cov_simple_het}
\alias{cov_simple_het}
\title{Compute covariance matrices with diagonal element 1 and
off-diagonal element corr}
\usage{
cov_simple_het(data, corr = c(0.25, 0.5, 0.75))
}
\arguments{
\item{data}{a mash data object, eg as created by \code{mash_set_data}}

\item{corr}{a vector containing the correlations to be used}
}
\value{
a list of matrices, one for each value in corr
}
\description{
This is an internal (non-exported) function. This help
  page provides additional documentation mainly intended for
  developers and expert users.
}
\keyword{internal}
