% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2cov.R
\name{data2cov_pca}
\alias{data2cov_pca}
\title{Perform PCA on data and return list of candidate covariance matrices}
\usage{
data2cov_pca(data, npc)
}
\arguments{
\item{data}{a mash data object}

\item{npc}{the number of PCs to use
Returns a list of covariance matrices: the k rank-one covariance matrices based on the first k PCs, and the
rank k covariance matrix}
}
\description{
Perform PCA on data and return list of candidate covariance matrices
}

