% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mash.R
\name{get_cov}
\alias{get_cov}
\title{Extract covariance matrices in m}
\usage{
get_cov(m)
}
\arguments{
\item{m}{a mash object}
}
\value{
a list of covariance matrices in m
}
\description{
Extract covariance matrices in m
}

