% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_functions.R
\name{get_cov}
\alias{get_cov}
\title{Return the covariances in g}
\usage{
get_cov(g)
}
\arguments{
\item{g}{a mixture of multivariate normals, as created for example by \code{initialize_g}}
}
\description{
Return the covariances in g
}

