% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_fitted_g}
\alias{get_fitted_g}
\title{Extract fitted g from mash object}
\usage{
get_fitted_g(m)
}
\arguments{
\item{m}{a mash object}
}
\value{
the fitted g (a list)
}
\description{
Extract fitted g from mash object
}
