% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mash.R
\name{get_posterior_matrices}
\alias{get_posterior_matrices}
\title{Return the posterior matrices from a mash object}
\usage{
get_posterior_matrices(m)
}
\arguments{
\item{m}{a mash object, as returned by \code{mash}}
}
\description{
Return the posterior matrices from a mash object
}

