% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods_common_cov.R
\name{is_common_cov}
\alias{is_common_cov}
\title{Check that all covariates are equal.}
\usage{
is_common_cov(data, Salpha = TRUE)
}
\arguments{
\item{data}{A mash data object.}

\item{Salpha}{Indicate the Shat to check (Shat or Shat_alpha)}
}
\description{
checks if all rows of Shat (or Shat_alpha) are the same - if so
    covariances are equal
}
