% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_covs.R
\name{make_names}
\alias{make_names}
\title{Create names for covariance matrices}
\usage{
make_names(names, suffixes)
}
\arguments{
\item{names}{a string}

\item{suffixes}{}
}
\description{
This is an internal (non-exported) function. This help
  page provides additional documentation mainly intended for
  developers and expert users.

Adds _suffixes to names for each element of suffixes
}
\keyword{internal}
