% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mash.R
\name{mash}
\alias{mash}
\title{Apply mash method to data}
\usage{
mash(data, Ulist = NULL, gridmult = sqrt(2), grid = NULL,
  normalizeU = TRUE, usepointmass = TRUE, g = NULL, fixg = FALSE,
  prior = c("nullbiased", "uniform"), optmethod = c("mixIP", "mixEM",
  "cxxMixSquarem"), verbose = TRUE, add.mem.profile = FALSE,
  algorithm.version = c("Rcpp", "R"), pi_thresh = 1e-10)
}
\arguments{
\item{data}{a mash data object containing the Bhat matrix and standard errors; created using \code{set_mash_data}}

\item{Ulist}{a list of covariance matrices to use}

\item{gridmult}{scalar indicating factor by which adjacent grid values should differ; close to 1 for fine grid}

\item{grid}{vector of grid values to use (scaling factors omega in paper)}

\item{normalizeU}{whether or not to normalize the U covariances to have maximum of 1 on diagonal}

\item{usepointmass}{whether to include a point mass at 0, corresponding to null in every condition}

\item{g}{the value of g obtained from a previous mash fit - an alternative to supplying Ulist, grid and usepointmass}

\item{fixg}{if g is supplied, allows the mixture proportions to be fixed rather than estimated - e.g. useful for fitting mash to test data after fitting it to training data}

\item{prior}{indicates what penalty to use on the likelihood, if any}

\item{optmethod}{name of optimization method to use}

\item{verbose}{If \code{TRUE}, print progress to R console.}

\item{add.mem.profile}{If \code{TRUE}, print memory usage to R console (requires R library `profmem`).}

\item{algorithm.version}{Indicates whether to use R or Rcpp version}

\item{pi_thresh}{threshold below which mixture components are ignored in computing posterior summaries (to speed calculations by ignoring negligible components)}
}
\value{
a list with elements result, loglik and fitted_g
}
\description{
Apply mash method to data
}
\examples{
Bhat = matrix(rnorm(100),ncol=5) # create some simulated data
Shat = matrix(rep(1,100),ncol=5)
data = mashr::set_mash_data(Bhat,Shat)
U.c = mashr::cov_canonical(data)
res.mash = mashr::mash(data,U.c)

}
