% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cov.R
\name{mash_add_cov}
\alias{mash_add_cov}
\title{Add a list of covariance matrices to m}
\usage{
mash_add_cov(m, cov_methods = c("identity", "singletons", "equal_effects",
  "simple_het"))
}
\arguments{
\item{m}{a mash object}

\item{Ulist}{a list of covariance matrices to add to m}
}
\description{
Add a list of covariance matrices to m
}
