% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cov.R
\name{mash_add_cov_ed}
\alias{mash_add_cov_ed}
\title{Add covariance matrices to m from "extreme deconvolution" (Bovy et al)}
\usage{
mash_add_cov_ed(m, Ulist_init = NULL, subset = NULL)
}
\arguments{
\item{m}{a mash object}

\item{Ulist_init}{a named list of covariance matrices to use to initialize ED; default is to use matrices from  PCs}

\item{subset}{a subset of data to be used when ED is run (default of NULL causes strong signals in m to be used)}
}
\description{
Add covariance matrices to m from "extreme deconvolution" (Bovy et al)
}
\details{
Runs the extreme deconvolution algorithm from Bovy et al (Annals of Applied Statistics) to estimate data-driven covariance matrices
The default is to initialize the EM algorithm from data2cov_pca with 5 PCs
}
