% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mash.R
\name{mash_calc_lik_matrix}
\alias{mash_calc_lik_matrix}
\title{Calculate the likelihood matrix for a mash object}
\usage{
mash_calc_lik_matrix(m)
}
\arguments{
\item{m}{the mash object}
}
\description{
Calculate the likelihood matrix for a mash object
}
\details{
Adds a J by P likelihood matrix to m, computed using the current data, grid and covariance matrices in m
}
