% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_mash.R
\name{mash_compute_loglik}
\alias{mash_compute_loglik}
\title{Compute loglikelihood for fitted mash object on new data}
\usage{
mash_compute_loglik(g, data)
}
\arguments{
\item{g}{a mash object or the fitted_g from a mash object}

\item{data}{a set of data on which to compute the loglikelihood}
}
\value{
the log-likelihood for data computed using g
}
\description{
Compute loglikelihood for fitted mash object on new data
}
