% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_mash.R
\name{mash_compute_posterior_matrices}
\alias{mash_compute_posterior_matrices}
\title{Compute posterior matrices for fitted mash object on new data}
\usage{
mash_compute_posterior_matrices(g, data)
}
\arguments{
\item{g}{a mash object or the fitted_g from a mash object}

\item{data}{a set of data on which to compute the posterior matrices}
}
\value{
A list of posterior matrices
}
\description{
Compute posterior matrices for fitted mash object on new data
}
