% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mash.R
\name{mash_compute_posterior_matrices}
\alias{mash_compute_posterior_matrices}
\title{Compute posterior matrices for fitted mash object on new data}
\usage{
mash_compute_posterior_matrices(g, data, pi_thresh = 1e-10,
  algorithm.version = c("Rcpp", "R"), A)
}
\arguments{
\item{g}{a mash object or the fitted_g from a mash object.}

\item{data}{a set of data on which to compute the posterior matrices}

\item{pi_thresh}{threshold below which mixture components are ignored in computing posterior summaries (to speed calculations by ignoring negligible components)}

\item{algorithm.version}{Indicates whether to use R or Rcpp version}

\item{A}{the linear transformation matrix}
}
\value{
A list of posterior matrices
}
\description{
Compute posterior matrices for fitted mash object on new data
}
