% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_functions.R
\name{optimize_g}
\alias{optimize_g}
\title{Fit the hierarchical model by estimating the mixture weights}
\usage{
optimize_g(data, g_init, prior = c("nullbiased"))
}
\arguments{
\item{data}{a mash data object, e.g. as created by \code{set_mash_data}}

\item{g_init}{a mixture of multivariate normals, e.g. as created by \code{initialize_g}}

\item{prior}{a string saying what kind of prior to use in the penalized likelihood}
}
\description{
Fit the hierarchical model by estimating the mixture weights
}

