% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_functions.R
\name{optimize_g}
\alias{optimize_g}
\title{Estimate the mixture weights for $g$ from a corresponding likelihood matrix}
\usage{
optimize_g(g_init, matrix_lik, prior = c("nullbiased"),
  optmethod = c("mixEM", "mixIP"), control = list())
}
\arguments{
\item{g_init}{a mixture of multivariate normals, e.g. as created by \code{add_to_g}}

\item{matrix_lik}{a matrix of likelihoods, where the (i,k)th entry is the probability of observation i given it came from component k of g}

\item{prior}{a string saying what kind of prior to use in the penalized likelihood}

\item{optmethod}{a string, giving name of optimization function to use}

\item{control}{a list of parameters to be passed to optmethod}
}
\value{
the optimal g, obtained by replacing the mixture proporions in g_init with the optimal mixture proportions
}
\description{
Estimate the mixture weights for $g$ from a corresponding likelihood matrix
}
