% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_data.R
\name{set_mash_data}
\alias{set_mash_data}
\title{Create a data object for mash analysis}
\usage{
set_mash_data(Bhat, Shat)
}
\arguments{
\item{Bhat}{an N by R matrix of observed estimates}

\item{Shat}{an N by R matrix of corresponding standard errors}
}
\value{
a data object for passing into mash functions
}
\description{
Create a data object for mash analysis
}
