% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_data.R
\name{set_mash_data}
\alias{set_mash_data}
\title{Create a data object for mash analysis}
\usage{
set_mash_data(Bhat, Shat, V = diag(ncol(Bhat)))
}
\arguments{
\item{Bhat}{an N by R matrix of observed estimates}

\item{Shat}{an N by R matrix of corresponding standard errors. Shat can be a scalar if all standard errors are equal.
This is most useful if Bhat is a matrix of Z scores, so elements of Shat are all 1.}

\item{V}{an R by R correlation matrix of error correlations; must be positive definite.
[So Bhat_j distributed as N(B_j,diag(Shat_j) V diag(Shat_j)) where _j denotes the jth row of a matrix]
Defaults to identity}
}
\value{
a data object for passing into mash functions
}
\description{
Create a data object for mash analysis
}
