% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations_contrast.R
\name{sim_contrast2}
\alias{sim_contrast2}
\title{Create simulation with signal data used for contrast
analysis.}
\usage{
sim_contrast2(nsamp = 1000, ncond = 5, err_sd = sqrt(0.5))
}
\arguments{
\item{nsamp}{Number of samples of each type.}

\item{ncond}{Number of conditions.}

\item{err_sd}{The standard deviation of the errors.}
}
\description{
Create simulation with signal data used for contrast
analysis.
}
\details{
The first condition is the reference group. The deviations
are the difference between the subsequent conditions with the
reference group.  The simulation consists of 90% null deviations,
10% non-nulls.  The non-nulls consist of equal numbers of three
different types of deviations: equal among conditions, present only
in the first subsequent condition, independent across conditions.
}
\examples{
sim_contrast2(100,5)

}
